move QueryServiceTest relating to construction of schema to its appropriate testing code
